Move the deprecation-related code
