+++ lib/routes/bundles.js
@@ -2,7 +2,7 @@
 
 const getStaticBundleStream = require('../../deprecated/static-bundle').getStaticBundleStream;
 const metrics = require('../monitoring/metrics');
-const handleDeprecatedRoute = require('../express/v1-deprecated');
+const handleDeprecatedRoute = require('../../deprecated/v1-deprecated');
 
 module.exports = function(app, config) {
 	const buildSystem = config.buildSystem;
