+++ lib/routes/docs.js
@@ -4,7 +4,7 @@ const fs = require('fs');
 const path = require('path');
 const oneWeek = 60 * 60 * 24 * 7;
 const hostnames = require('../utils/hostnames');
-const handleDeprecatedRoute = require('../express/v1-deprecated');
+const handleDeprecatedRoute = require('../../deprecated/v1-deprecated');
 
 module.exports = function(app, config) {
