+++ lib/routes/modules.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const createRouteHandler = require('../express/promisehandler');
-const handleDeprecatedRoute = require('../express/v1-deprecated');
+const handleDeprecatedRoute = require('../../deprecated/v1-deprecated');
 
 module.exports = function(app, config) {
 	const buildSystem = config.buildSystem;
